requirements.txt
ultralytics 
opencv-python 
roboflow
matplotlib
pillow 
numpy 
pandas 
PyYAML 
imagehash 
fastapi
uvicorn
python-multipart

# torch >= 2.0.0
# torchvision >= 0.15.0